+++ src/Kit/DataDisplay/IdCard/types.d.ts
@@ -1,0 +1,9 @@
+import {KitHTMLAttributes} from '../../../types';
+import {ReactNode} from 'react';
+
+export interface IKitIdCard extends KitHTMLAttributes<HTMLDivElement> {
+    avatar?: ReactNode;
+    size?: 'medium' | 'large';
+    title?: string;
+    description?: string;
+}