+++ stories/02-Design System/05-DataDisplay/IdCard/examples/basic.tsx
@@ -1,0 +1,8 @@
+import {KitIdCard} from '@kit/DataDisplay';
+import React from 'react';
+
+const App = () => {
+    return <KitIdCard title="John Doe" description="A talented software developer." />;
+};
+
+export default App;