+++ stories/02-Design System/05-DataDisplay/IdCard/examples/index.tsx
@@ -1,0 +1,19 @@
+import Basic from './basic';
+import Avatar from './avatar';
+import Size from './size';
+
+import BasicSource from './basic?raw';
+import AvatarSource from './avatar?raw';
+import SizeSource from './size?raw';
+
+export const Sources = {
+    Basic: BasicSource,
+    Avatar: AvatarSource,
+    Size: SizeSource
+};
+
+export default {
+    Basic,
+    Avatar,
+    Size
+};