+++ stories/index.tsx
@@ -27,6 +27,7 @@ import {EditorTemplate as Badge} from './02-Design System/05-DataDisplay/Badge/T
 import {EditorTemplate as Card} from './02-Design System/05-DataDisplay/Card/Template';
 import {EditorTemplate as Collapse} from './02-Design System/05-DataDisplay/Collapse/Template';
 import {EditorTemplate as Empty} from './02-Design System/05-DataDisplay/Empty/Template';
+import {EditorTemplate as IdCard} from './02-Design System/05-DataDisplay/IdCard/Template';
 import {EditorTemplate as Image} from './02-Design System/05-DataDisplay/Image/Template';
 import {EditorTemplate as ItemCard} from './02-Design System/05-DataDisplay/ItemCard/Template';
 import {EditorTemplate as ItemList} from './02-Design System/05-DataDisplay/ItemList/Template';
@@ -78,6 +79,7 @@ export default {
         Card,
         Collapse,
         Empty,
+        IdCard,
         Image,
         ItemCard,
         ItemList,